+++ lib/playground_web/live/home_live.ex
@@ -11,10 +11,11 @@ defmodule PlaygroundWeb.HomeLive do
     {:noreply, apply_action(socket, socket.assigns.live_action, params)}
   end
 
-  defp apply_action(socket, :index, _params) do
+  defp apply_action(socket, :index, params) do
     socket
     |> assign(:page_title, "Welcome!")
     |> assign(:room, nil)
+    |> maybe_put_flash(params)
   end
 
   defp apply_action(socket, :new, _params) do
@@ -27,5 +28,16 @@ defmodule PlaygroundWeb.HomeLive do
     socket
     |> assign(:page_title, "Join Room")
     |> assign(:room, %{"id" => nil, "room_code" => room_code, "host_name" => ""})
+  end
+
+  defp maybe_put_flash(socket, %{"error" => error}) do
+    # A bit more safe
+    sliced_error = String.slice(error, 0..120)
+
+    put_flash(socket, :error, sliced_error)
+  end
+
+  defp maybe_put_flash(socket, _params) do
+    socket
   end
 end